REGRESSION (271599@main): espn.com: Closed captions aren't displaying in videos https://bugs.webkit.org/show_bug.cgi?id=269847 rdar://122842659
Reviewed by Eric Carlson.

In this video, we had 2 text tracks, all with a track id of 0.
The reasons for this is that it contained a InbandMetadataTextTrack and a InbandTextTrack
and the MediaPlayerPrivateAVFoundation weren't assigning them unique trackID which
would confuse the search in the hash table trying to match a TrackID to a track.

A TrackID for an inband text track is an abstract concept, it doesn't reflect anything
relevant to the user (Except that it must be unique).
We only need that TrackID to be unique between all the tracks in a given MediaPlayer.

So we assign them a unique TrackID at creation.

Add a HLS file that contains both a metadata and a subtitle track and ensure
that cues are added to the proper one.

* LayoutTests/http/tests/media/hls/track-webvtt-multitracks-expected.txt: Added.
* LayoutTests/http/tests/media/hls/track-webvtt-multitracks.html: Added.
* LayoutTests/http/tests/media/resources/hls/test-webvtt-multitracks.m3u8: Added.
* Source/WebCore/platform/graphics/avfoundation/InbandMetadataTextTrackPrivateAVF.cpp:
(WebCore::InbandMetadataTextTrackPrivateAVF::create):
(WebCore::InbandMetadataTextTrackPrivateAVF::InbandMetadataTextTrackPrivateAVF):
* Source/WebCore/platform/graphics/avfoundation/InbandMetadataTextTrackPrivateAVF.h:
* Source/WebCore/platform/graphics/avfoundation/InbandTextTrackPrivateAVF.cpp:
(WebCore::InbandTextTrackPrivateAVF::InbandTextTrackPrivateAVF):
* Source/WebCore/platform/graphics/avfoundation/InbandTextTrackPrivateAVF.h:
* Source/WebCore/platform/graphics/avfoundation/objc/InbandTextTrackPrivateAVFObjC.h:
(WebCore::InbandTextTrackPrivateAVFObjC::create):
* Source/WebCore/platform/graphics/avfoundation/objc/InbandTextTrackPrivateAVFObjC.mm:
(WebCore::InbandTextTrackPrivateAVFObjC::InbandTextTrackPrivateAVFObjC):
(WebCore::InbandTextTrackPrivateAVFObjC::id const): Deleted.
* Source/WebCore/platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::MediaPlayerPrivateAVFoundationObjC::processMediaSelectionOptions):
(WebCore::MediaPlayerPrivateAVFoundationObjC::processMetadataTrack):
* Source/WebCore/platform/graphics/avfoundation/objc/OutOfBandTextTrackPrivateAVF.h:
(WebCore::OutOfBandTextTrackPrivateAVF::create):
(WebCore::OutOfBandTextTrackPrivateAVF::OutOfBandTextTrackPrivateAVF):

Canonical link: https://commits.webkit.org/275201@main