[WPE] Enable autoplay and WebRTC in MiniBrowser https://bugs.webkit.org/show_bug.cgi?id=269920
Reviewed by Michael Catanzaro.

Before 264689@main, the autoplay setting was not properly applied and
was always WEBKIT_AUTOPLAY_ALLOW for WPE. Now, the default setting of
WEBKIT_AUTOPLAY_ALLOW_WITHOUT_SOUND is really applied.

This prevents some pages with minimal UI from playing videos, and
MiniBrowser does not have the UI to enable autoplay.
It is easier to enable it by default.

Also enable WebRTC, which has been made opt-in in 251951@main, and now
needs to be enabled in addition to MediaStream.

* Tools/MiniBrowser/wpe/main.cpp:
(activate):

Canonical link: https://commits.webkit.org/275241@main