[css-grid] Turn masonry feature off-by-default on non-Cocoa builds https://bugs.webkit.org/show_bug.cgi?id=269939 <rdar://problem/123465041>
Reviewed by Tim Nguyen.

In Bug 247074 we moved the Masonry feature flag from preview to stable. While
the Cocoa build of WebKit needs the feature to be on by default at the moment,
we do not need to toggle the feature for other ports.

This patch leaves the Cocoa port enabled, but switches the default for non-Cocoa
ports back off.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:

Canonical link: https://commits.webkit.org/275249@main