[MSE] Media from the TimeRange containing currentTime can be evicted when it shouldn't https://bugs.webkit.org/show_bug.cgi?id=270034 rdar://problem/123556056
Reviewed by Eric Carlson.

We want none of the data contiguous from the currentTime to the next discontinuity to be
evicted to ensure continuous playback.
This is inline with Google's proposal for having Eviction Policies.

The comments in the eviction code indicated as such, but a logic error would have made
the end of the current range be truncated if it was followed by another range
after a discontinuity.

* LayoutTests/media/media-source/media-managedmse-eviction-expected.txt: Added.
* LayoutTests/media/media-source/media-managedmse-eviction.html: Added.
* LayoutTests/platform/glib/TestExpectations:
* Source/WebCore/platform/graphics/SourceBufferPrivate.cpp:
(WebCore::SourceBufferPrivate::evictFrames):

Canonical link: https://commits.webkit.org/275306@main