View transition layers are positioned wrong if a clipping layer is present. https://bugs.webkit.org/show_bug.cgi?id=270058 <rdar://123607249>
Reviewed by Tim Nguyen.

Positioning for a view-transition 'new' layer currently sets the origin of the
primary graphics rect to 0,0, but this doesn't move any clipping layers positioned
in updateClippingStackLayerGeometry.

We can instead adjust the position of the 'parent graphics layer rect' (since all
local layer positions are computed relative to that) to a value that results in 0,0
being computed for the primary rect.

* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::updateGeometry):

Canonical link: https://commits.webkit.org/275307@main