instanceof should not get RHS prototype when LHS is primitive https://bugs.webkit.org/show_bug.cgi?id=270065
Reviewed by Justin Michaud.

The expression `x instanceof obj` should not access obj.prototype when x is primitive per the spec,
but our implementation eagerly grabs the prototype in order to hand it off to OpInstanceof.

    https://tc39.es/ecma262/multipage/abstract-operations.html#sec-ordinaryhasinstance
    7.3.21 OrdinaryHasInstance ( C, O )
        ...
        3. If O is not an Object, return false.
        4. Let P be ? Get(C, "prototype").
        ...

We could refactor OpInstanceof to take the RHS directly instead of its prototype, but it suffices to add a couple of
lines to InstanceOfNode::emitBytecode. (And while we're at it, we can also do a bit better with temp register reuse.)

* JSTests/test262/expectations.yaml: Mark two tests as passing.
* Source/JavaScriptCore/bytecompiler/NodesCodegen.cpp:
(JSC::InstanceOfNode::emitBytecode):

Canonical link: https://commits.webkit.org/275318@main