[UnifiedPDF] PluginView::contentScaleFactor reports an incorrect content scale factor https://bugs.webkit.org/show_bug.cgi?id=270077 rdar://123624248
Reviewed by Tim Horton.

This method plumbs out the result of the plugin's scaleFactor() method,
and not of its namesake. This patch fixes that.

This patch thus addresses the wrongly sized lookup highlights on force
click, which were ill-sized because we generated the wrong TI data in
WebPage::dictionaryPopupInfoForSelectionInPDFPlugin.

* Source/WebKit/WebProcess/Plugins/PluginView.cpp:
(WebKit::PluginView::contentScaleFactor const):

Canonical link: https://commits.webkit.org/275347@main