Preload restricted window opener data https://bugs.webkit.org/show_bug.cgi?id=270106 rdar://123640341
Reviewed by Wenson Hsieh.

The first call to check for the restricted opener type for a domain currently always claims that the
domain should have an unrestricted opener. This is because we load the list of restricted openers
from WebPrivacy both lazily and asynchronously.

To make it much more likely that the first call has up-to-date data, we now load the data in
`WebProcessPool::didReachGoodTimeToPrewarm`.

* Source/WebKit/UIProcess/WebProcessPool.cpp:
(WebKit::loadRestrictedOpenerTypeDataIfNeeded):
(WebKit::WebProcessPool::didReachGoodTimeToPrewarm):

Canonical link: https://commits.webkit.org/275365@main