[UnifiedPDF] Word selections should not be generated when clicking on annotations https://bugs.webkit.org/show_bug.cgi?id=270140 rdar://123666112
Reviewed by Tim Horton.

We are unnecessarily overriding the current selection in
lookupTextAtLocation(). This manifests itself in the creation of stray
word selections when clicking on annotations, and also in the brief
flash of a blue highlight over words when simply clicking on them.

The former is more apparent because we don't immediately reset the
erroneously set selection with an empty one, but both are a result of
WebPage calling into lookupTextAtLocation() as part of immediate action
hit testing.

This patch addresses the issue by simply not overriding the tracked
selection from within lookupTextAtLocation().

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::lookupTextAtLocation):

Canonical link: https://commits.webkit.org/275370@main