Web Inspector: Font Panel: CSS font property values marked !important don't get overridden when using the interactive editing controls. https://bugs.webkit.org/show_bug.cgi?id=258975
Reviewed by Devin Rousso.

If a CSS class has a !important tag on a property the slider does not work correctly.
This is due to us modifying the inline style, and the inline style is being overriden by
the css class !important property.

We can address this by checking if the !important property is being applied already
during _computeProperty(), and then if it is then append !important in the
writeFontVariation().

Rename resultProperties to resultProperty inside of _computeProperty() because there
is a single property being handled.

Delete _effectivePropertyValueForName function since it is not in use and create it in
_computeProperty instead to be used for resultProperty.isImportant.

* Source/WebInspectorUI/UserInterface/Models/FontStyles.js:
(WI.FontStyles.prototype.writeFontVariation):
(WI.FontStyles):

Canonical link: https://commits.webkit.org/275389@main