[UnifiedPDF] Don't show the form hover highlight for non-fillable forms https://bugs.webkit.org/show_bug.cgi?id=270184 rdar://123667854
Reviewed by Tim Horton.

If the plugin does not support forms, because it is not a main frame
or iframe plugin, then we should not start tracking the text widget
annotations. This will make sure that we do not keep around any state
that would be used later on for painting the hover indicator.

* Source/WebKit/WebProcess/Plugins/PDF/UnifiedPDF/UnifiedPDFPlugin.mm:
(WebKit::UnifiedPDFPlugin::paintHoveredAnnotationOnPage):
(WebKit::UnifiedPDFPlugin::handleMouseEvent):

Canonical link: https://commits.webkit.org/275413@main