[UnifiedPDF] Drop support for "Search in Spotlight" context menu item https://bugs.webkit.org/show_bug.cgi?id=269320 rdar://122911045
Reviewed by Tim Horton.

This hasn't worked in a while (cf. rdar://56416395), is missing in
PDFView-based context menu items, and its value is marginal. Let's drop
the plumbing so it's squarely a no-op for now.

Note that we can't get rid of the delegate method implementation since
PDFLayerController calls into it unconditionally, but that can be fixed
at a later point.

* Source/WebKit/WebProcess/Plugins/PDF/PDFPlugin.mm:
(-[WKPDFLayerControllerDelegate performSpotlightSearch:]):
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginBase.h:
* Source/WebKit/WebProcess/Plugins/PDF/PDFPluginBase.mm:
(WebKit::PDFPluginBase::performSpotlightSearch): Deleted.

Canonical link: https://commits.webkit.org/275426@main