Safari unable to play YouTube content fast enough. Content window spins and plays only after 25-30 seconds https://bugs.webkit.org/show_bug.cgi?id=270205 rdar://123528095
Reviewed by Eric Carlson.

Both MSE and MediaSourceRemote and RemoteMediaResourceLoader WorkQueue were created using QOS::Background priority.
Trace shows that when a task is being run on those WorkQueues, they are either being pre-empted or yielded and sometimes
it takes over 10s for a task to even complete.

We change the priority to the default, the same as what the IPC's WorkQueue is using.

* Source/WebKit/GPUProcess/media/RemoteMediaResourceLoader.h:
* Source/WebKit/WebProcess/GPU/media/MediaSourcePrivateRemote.cpp:
(WebKit::MediaSourcePrivateRemote::queue):

Canonical link: https://commits.webkit.org/275443@main